+++ tests/index.ts
@@ -4,5 +4,6 @@ import 'reflect-metadata';
 
 // tests
 import './angular2Apollo';
+import './apolloQueryObservable';
 import './apolloQueryPipe';
 import './apolloDecorator/index';